rework if statements
Rather than requiring ';' to signify then end of the condition and 'then' block
if simply treats the next token as the condition and the token after that then 'then' block

This simplifies simple if expressions such as "if x; y;" to "if x y".
Hence they will be easier to use in expression, once 'else' is implemented.
e.g.: "x = if y z else 0"

expressions as conditions can be acieved by the use of parentheses, the same goes for 'then' blocks
e.g. "if (x+2) (x+10)" will add ten to x if x is not -2