add initial support for builtin functions
builtin functions are a good way to extend the language for application specific purposes.
They are denoted by ':' to differentiate them from variables (for now).
Note that while they are currently treated as keywords the idea is to be able to pass in
some sort of data structure to specify which builtin functions are supported for a
particular use case.