fix jumps to end of progam causing Exception
Jumping to the end of a program caused IndexOutOfBounds Exceptions since the
label would be replaced by the address of the next instruction. However, if the
label is the last instruction, the 'next' instruction is outside of the buffer.

This can happen when the program neds with an 'if' block, such as: "5 + if 1 2"
The fix is to insert a NOP (no operation) at the end of the program if the last
ByteCode is a label. This causes the address generated by the label to be legal
and therefore no exceptions occur.